pyproject.toml
[tool.poetry]
name = "atm"
version = "1.0.0"
description = "mlops"
authors = ["Semyon-Yakovlev <patt1337gg@gmail.com>",
           "blarno <bigsquare@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0.1"
toml = "^0.10.2"
numpy = "1.22.3"
matplotlib = "^3.5.1"
torch = "2.1.0"
pandas = "1.2.2"
fire = "^0.5.0"
joblib = "^1.3.2"
hydra-core = "^1.3.2"
mlflow = "^1.14.0"
torcheval = "^0.0.7"
scikit-learn = "^1.3.2"
catboost = "^1.2.2"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
api = "^0.0.7"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
isort = "^5.12.0"
dvc = "^3.36.1"
dvc-gdrive= "0.0.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"